Ai/requirements.txt
sounddevice
numpy
tensorflow
pyserial
librosa
